+++ src/views/AView.ts
@@ -7,7 +7,7 @@ import {EventHandler} from 'phovea_core/src/event';
 import {defaultSelectionType, IDType, resolve} from 'phovea_core/src/idtype';
 import {none} from 'phovea_core/src/range';
 import {IFormElementDesc, FormBuilder} from '../form';
-import {toData} from '../form/internal/AFormElement';
+import {toData} from '../form/elements/AFormElement';
 import {
   EViewMode, ISelection, isSameSelection, IView, IViewContext, VIEW_EVENT_ITEM_SELECT,
   VIEW_EVENT_LOADING_FINISHED, VIEW_EVENT_UPDATE_ENTRY_POINT, VIEW_EVENT_UPDATE_SHARED
